+++ apps/web-nuha/components/atoms/index.ts
@@ -3,3 +3,4 @@ export * from './Ring';
 export * from './ProgressBar';
 export * from './Kosong';
 export * from './Avatar';
+export * from './IkonOpsi';